refactor hit search kernels to make space for cuda versions
