+++ .gitignore
@@ -6,3 +6,4 @@ notebooks/
 .vscode/
 .py-build-cmake_cache/
 temp/
+dist/
