+++ bland/bland/include/bland/stride_helper.hpp
@@ -1,0 +1,28 @@
+#pragma once
+
+#include <cstdint>
+#include <vector>
+
+namespace bland {
+
+using nd_coords = std::vector<int64_t>;
+
+// Helper to abstract out increments (this is copied in local_maxima, TODO: factor out to bland utils)
+struct stride_helper {
+
+    std::vector<int64_t> shape;
+    std::vector<int64_t> stride;
+
+    stride_helper(std::vector<int64_t> shape, std::vector<int64_t> stride) : shape(shape), stride(stride) {}
+
+    int64_t to_linear_offset(const nd_coords &coords) const {
+        // TODO compute linear index of curr_coords from shape and stride
+        auto linear_offset = 0;
+        for (int dim = 0; dim < shape.size(); ++dim) {
+            linear_offset += coords[dim] * stride[dim]; // TODO: also add offset
+        }
+        return linear_offset;
+    }
+};
+
+}
