+++ bliss/drift_search/kernels/connected_components_cpu.hpp
@@ -1,0 +1,28 @@
+#pragma once
+
+#include <bland/ndarray.hpp>
+
+#include <drift_search/hit_search.hpp> // component
+
+#include <vector>
+
+namespace bliss {
+
+/**
+ * find clusters (components) of adjacent (in start frequency or drift rate) bins and group them together.
+ *
+ * Accepts a binary mask (1) of dtype uint8
+ */
+std::vector<component> find_components_in_binary_mask_cpu(const bland::ndarray         &threshold_mask,
+                                                          std::vector<bland::nd_coords> neighborhood);
+
+/**
+ * Given noise stats do a combined threshold and cluster of nearby components
+ */
+std::vector<component>
+find_components_above_threshold_cpu(bland::ndarray                       doppler_spectrum,
+                                    integrated_flags                     dedrifted_rfi,
+                                    std::vector<std::pair<float, float>> noise_and_thresholds_per_drift,
+                                    std::vector<bland::nd_coords>        max_neighborhood);
+
+} // namespace bliss
