+++ bliss/drift_search/kernels/local_maxima_cuda.cu
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "local_maxima_cu.h"
+
+using namespace bliss;
+
+std::vector<component> bliss::find_local_maxima_above_threshold_cuda(coarse_channel        &dedrifted_coarse_channel,
+                                                         float                  snr_threshold,
+                                                         std::vector<nd_coords> max_neighborhoods) {
+
+    return {};
+}
