+++ bliss/drift_search/kernels/local_maxima_cuda.cuh
@@ -1,10 +1,10 @@
 #pragma once
 
-#include "hit_search.hpp"
+#include <drift_search/hit_search.hpp> // component
 
 namespace bliss {
 
-std::vector<component> find_local_maxima_above_threshold(coarse_channel        &dedrifted_coarse_channel,
+std::vector<component> find_local_maxima_above_threshold_cuda(coarse_channel        &dedrifted_coarse_channel,
                                                          float                  snr_threshold,
                                                          std::vector<bland::nd_coords> max_neighborhoods);
